gh website setup
This commit puts all source material in the source directory,
and introduces the ``make docs`` command to create a html website
in the ``docs`` directory. The ``make clean-cache`` will clear
the website material and leave the ``docs`` folder with only
the ``source`` folder. Intended to facilitate transition to
Github Pages.