Fixes for refit option, learner cloning, lyr push
Fitting with refit False caused errors for several reasons.

  - Changing the outer fit call to transform broke the sublearner
    job call. Sublearners now register the job in the constructor
    to allow the learner to dynamically change the job type if
    refit is false

  - Collect was always called, but transform does not cache
    sublearners, so the learner ended up dropping its fitted
    sublearners. A new __collect__ call is introduced in the
    _BaseEstimator class. It should be turned on only in a
    true gen_fit call, and is then turned off after collect.

Cloning with auxiliary was broken as the auxiliary was not cloned.
Solution is to explicitly introduce the auxiliary as a parameter.
The learner constructor will force-tie the indexer of the
auxiliary with its own.

Layer now allows multiple groups in push.